+++ axelor-stock/src/main/java/com/axelor/apps/stock/service/StockMoveServiceImpl.java
@@ -474,16 +474,14 @@ public class StockMoveServiceImpl implements StockMoveService {
 
     try {
       if (stockMove.getIsWithBackorder() && mustBeSplit(stockMove.getStockMoveLineList())) {
-        Optional<StockMove> newStockMove =
-            copyAndSplitStockMove(stockMove, stockMove.getPlannedStockMoveLineList());
+        Optional<StockMove> newStockMove = copyAndSplitStockMove(stockMove);
         if (newStockMove.isPresent()) {
           newStockSeq = newStockMove.get().getStockMoveSeq();
         }
       }
 
       if (stockMove.getIsWithReturnSurplus() && mustBeSplit(stockMove.getStockMoveLineList())) {
-        Optional<StockMove> newStockMove =
-            copyAndSplitStockMoveReverse(stockMove, stockMove.getPlannedStockMoveLineList(), true);
+        Optional<StockMove> newStockMove = copyAndSplitStockMoveReverse(stockMove, true);
         if (newStockMove.isPresent()) {
           if (newStockSeq != null) {
             newStockSeq = newStockSeq + " " + newStockMove.get().getStockMoveSeq();